+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Hook useSaveWithGlobalRewrite() ([#161](https://github.com/i-VRESSE/workflow-builder/pull/161))
 * Support for if/then/else in JSON schema ([#160](https://github.com/i-VRESSE/workflow-builder/pull/160))
+* Pick any chain of the given molecules ([#162](https://github.com/i-VRESSE/workflow-builder/issues/162))
 
 ## @i-vresse/wb-core 3.0.1 - 2024-05-23